+++ src/commands/text/fuck_tyler.ts
